requirements.txt
Django
django_countries
django-rest-framework
social-auth-app-django
django-cors-headers
psycopg2
geoip2
captcha
django-import-export
python-dateutil
django-richtextfield
django-mathfilters
inflect
pyyaml
ua-parser
user-agents
django-user-agents
qrcode[pil]
django-translations
googletrans